Add shared DB retry logic
